+++ server/src/events/handlers/action_closed_event_handler.py
@@ -3,14 +3,22 @@ This module implements the ActionClosedEventHandler class.
 """
 from typing import List
 
-from src import Event, Notification, MemberRole
+from src import Event, Notification, MemberRole, get_settings
 from src.database import SessionLocal
 from src.events.handlers.event_handler import EventHandler
+from src.utils.backup import PostgresBackupManager
 from src.utils.enums import NotificationType, Role
 
 
 class ActionClosedEventHandler(EventHandler):
     def handle(self, event: Event) -> None:
+        try:
+            self._send_database_backup_email(event=event)
+        except Exception as e:
+            self._logger.error(
+                f"An error occurred while sending the database backup email: {e}"
+            )
+
         session = SessionLocal()
 
         observer_member_roles: List[MemberRole] = (
@@ -38,3 +46,17 @@ class ActionClosedEventHandler(EventHandler):
             session.add(notification)
 
         session.commit()
+
+    def _send_database_backup_email(self) -> None:
+        """
+        Send an email with the database backup attached.
+        :return: None
+        """
+        settings = get_settings()
+        recipient = settings.database_backup_recipient_email
+
+        self._logger.info(f"Sending database backup email to {recipient}")
+        backup_manager = PostgresBackupManager.from_env()
+        backup_manager.backup_and_send_to_recipient()
+
+        self._logger.info(f"Database backup email sent to {recipient}")